ANOTHER round of type fixes.
Fix ulong -> uint32_t.

We keep hitting these but we're very close to having libc building ...

Somebody needs to figure out what to do about USED. It can take many parameterswq

+++ 386/include/u.h
@@ -54,8 +54,8 @@ union FPdbleword
 {
 	double	x;
 	struct {	/* little endian */
-		ulong lo;
-		ulong hi;
+		uint32_t lo;
+		uint32_t hi;
 	};
 };
